+++ helium-dev/src/main/java/org/apache/zeppelin/helium/DevInterpreter.java
@@ -14,7 +14,8 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.zeppelin.interpreter.dev;
+
+package org.apache.zeppelin.helium;
 
 import java.util.LinkedList;
 import java.util.List;
@@ -48,7 +49,7 @@ public class DevInterpreter extends Interpreter {
   }
 
   /**
-   * event handler for ZeppelinApplicationDevServer
+   * event handler for org.apache.zeppelin.helium.ZeppelinApplicationDevServer
    */
   public static interface InterpreterEvent {
     public InterpreterResult interpret(String st, InterpreterContext context);
