+++ zeppelin-interpreter/src/main/java/org/apache/zeppelin/interpreter/remote/RemoteInterpreterServer.java
@@ -17,16 +17,12 @@
 
 package org.apache.zeppelin.interpreter.remote;
 
-
 import java.io.IOException;
 import java.lang.reflect.Constructor;
 import java.lang.reflect.InvocationTargetException;
 import java.net.URL;
 import java.nio.ByteBuffer;
-import java.rmi.server.RemoteServer;
 import java.util.*;
-import java.util.concurrent.ConcurrentHashMap;
-import java.util.concurrent.TimeoutException;
 
 import org.apache.thrift.TException;
 import org.apache.thrift.server.TThreadPoolServer;
@@ -39,7 +35,7 @@ import org.apache.zeppelin.interpreter.*;
 import org.apache.zeppelin.interpreter.InterpreterHookRegistry.HookType;
 import org.apache.zeppelin.interpreter.InterpreterHookListener;
 import org.apache.zeppelin.interpreter.InterpreterResult.Code;
-import org.apache.zeppelin.interpreter.dev.ZeppelinDevServer;
+import org.apache.zeppelin.helium.ZeppelinDevServer;
 import org.apache.zeppelin.interpreter.thrift.*;
 import org.apache.zeppelin.resource.*;
 import org.apache.zeppelin.scheduler.Job;
@@ -442,7 +438,7 @@ public class RemoteInterpreterServer
         public void onPreExecute(String script) {
           String cmdDev = interpreter.getHook(noteId, HookType.PRE_EXEC_DEV);
           String cmdUser = interpreter.getHook(noteId, HookType.PRE_EXEC);
-          
+
           // User defined hook should be executed before dev hook
           List<String> cmds = Arrays.asList(cmdDev, cmdUser);
           for (String cmd : cmds) {
@@ -450,15 +446,15 @@ public class RemoteInterpreterServer
               script = cmd + '\n' + script;
             }
           }
-          
+
           InterpretJob.this.script = script;
         }
-        
+
         @Override
         public void onPostExecute(String script) {
           String cmdDev = interpreter.getHook(noteId, HookType.POST_EXEC_DEV);
           String cmdUser = interpreter.getHook(noteId, HookType.POST_EXEC);
-          
+
           // User defined hook should be executed after dev hook
           List<String> cmds = Arrays.asList(cmdUser, cmdDev);
           for (String cmd : cmds) {
@@ -466,7 +462,7 @@ public class RemoteInterpreterServer
               script += '\n' + cmd;
             }
           }
-          
+
           InterpretJob.this.script = script;
         }
       };
@@ -478,7 +474,7 @@ public class RemoteInterpreterServer
     protected Object jobRun() throws Throwable {
       try {
         InterpreterContext.set(context);
-        
+
         // Open the interpreter instance prior to calling interpret().
         // This is necessary because the earliest we can register a hook
         // is from within the open() method.
@@ -486,7 +482,7 @@ public class RemoteInterpreterServer
         if (!lazy.isOpen()) {
           lazy.open();
         }
-        
+
         // Add hooks to script from registry.
         // Global scope first, followed by notebook scope
         processInterpreterHooks(null);
